fix: NavigateButton 기능 분리
