+++ src/pages/event/WriteEventInfoPage/WriteEventInfoPage.style.ts
@@ -40,43 +40,9 @@ const TwoInputContainer = styled.div`
 const HalfInputForm = styled(InputForm)`
   width: 50%;
 `;
-const ButtonWrapper = styled.div`
-  display: flex;
-  position: absolute;
-  right: 0;
-  bottom: -2rem;
-  gap: 1rem;
-  padding: 1rem 0;
-`;
-const Button = styled.button`
-  width: 6rem;
-  height: 3rem;
-  padding: 0.5rem;
-  box-sizing: border-box;
-  border: none;
-  border-radius: 0.25rem;
-  outline: none;
-  cursor: pointer;
-  color: white;
-`;
-const PrevButton = styled(Button)`
-  background-color: ${Theme.color.idkGrey};
-`;
-const SubmitButton = styled(Button)`
-  background-color: ${Theme.color.tSemiActive};
-`;
 const ErrorMessage = styled.p`
   font-size: 0.7rem;
   color: red;
 `;
 
-export {
-  PerformanceFormContainer,
-  ContentArea,
-  TwoInputContainer,
-  HalfInputForm,
-  ButtonWrapper,
-  PrevButton,
-  SubmitButton,
-  ErrorMessage,
-};
+export { PerformanceFormContainer, ContentArea, TwoInputContainer, HalfInputForm, ErrorMessage };
